:wrench: Add :wrench: to package.json and skip config creation if already exists.

+++ scripts/createConfig.js
@@ -1,12 +1,16 @@
-const fs = require('fs')
-const path = require('path')
+import fs from 'fs'
+import path from 'path'
 
 const homeDir =
   process.env[process.platform === 'win32' ? 'USERPROFILE' : 'HOME']
 const configDir = path.join(homeDir, '.config', 'gpt-cli')
 const configFile = path.join(configDir, '.gptrc')
 
-// 创建配置文件夹和配置文件
+// if exits, do nothing, process exits with code 0
+if (fs.existsSync(configFile)) {
+  process.exit(0)
+}
+
 fs.mkdirSync(configDir, { recursive: true })
 fs.writeFileSync(configFile, '', { mode: 0o600 })
